// data/github/current/ui/packages/sub-issues/components/__generated__/SubIssuesListItem.graphql.ts
export type SubIssuesListItem$data = {
  readonly assignees: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly avatarUrl: string;
        readonly id: string;
        readonly login: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly id: string;
  readonly repository: {
    readonly name: string;
    readonly owner: {
      readonly login: string;
    };
  };
  readonly url: string;
  readonly " $fragmentSpreads": FragmentRefs<"ClosedByPullRequestsReferences" | "SubIssueStateIcon" | "SubIssuesListItem_NestedSubIssues">;
  readonly " $fragmentType": "SubIssuesListItem";
};
export type SubIssuesListItem$key = {
  readonly " $data"?: SubIssuesListItem$data;
  readonly " $fragmentSpreads": FragmentRefs<"SubIssuesListItem">;
};